Heavy Update
No extraneous scripts. All name conventions are the same. Everything should be linked well. Added UI for adding and deleting contacts. contact_manager is successfuly querying and displaying database information. Last things that need to be implemented are the php files for doing the inserting and deleting into the database for the contacts, and a search bar.